fix(agy): address review feedback and isolate Gemini translation
