+++ .github/workflows/ci.yml
@@ -859,13 +859,59 @@ jobs:
         run: |
           & ([scriptblock]::Create((irm https://hermes-agent.nousresearch.com/install.ps1))) -SkipSetup
 
-      # AGY is an optional native CLI. The pinned official release and its
-      # per-platform SHA-256 checksums live in mise.toml/mise.lock; Renovate
-      # should update those files together.
-      - name: Install AGY with mise
-        uses: jdx/mise-action@7e36c90d9ab29c415a2384db3006f3ec8a8cc654 # v4.2.4
-        with:
-          github_token: ${{ github.token }}
+      # Pinned official AGY release; verify each runner's archive before installing.
+      - name: Install AGY (Unix)
+        if: runner.os != 'Windows'
+        shell: bash
+        env:
+          AGY_VERSION: 1.1.13
+          AGY_PLATFORM: ${{ runner.os }}-${{ runner.arch }}
+        run: |
+          set -euo pipefail
+          case "$AGY_PLATFORM" in
+            Linux-X64) archive=agy_cli_linux_x64.tar.gz; checksum=edc7c32b5ab4fc2e4da03381fee83ed566dea6b56b56f9329cd13cd77947a1d9 ;;
+            Linux-ARM64) archive=agy_cli_linux_arm64.tar.gz; checksum=a9fdd2a386770c27dbf784436bd4de70d4d4901c832d5ec6abf27758d5c370f8 ;;
+            macOS-X64) archive=agy_cli_mac_x64.tar.gz; checksum=8eaf399051afbc67435220f7fbeddeaaa6f9d50a087a229e628a4b2510c05c69 ;;
+            macOS-ARM64) archive=agy_cli_mac_arm64.tar.gz; checksum=f8c27b70708843eeb2c8f6efff0119a27041980e95cba6c49b86a0950ed19ba1 ;;
+            *) echo "Unsupported AGY platform: $AGY_PLATFORM" >&2; exit 1 ;;
+          esac
+          install_dir="$RUNNER_TEMP/agy"
+          mkdir -p "$install_dir"
+          curl --fail --location --retry 3 "https://github.com/google-antigravity/antigravity-cli/releases/download/$AGY_VERSION/$archive" -o "$install_dir/$archive"
+          cd "$install_dir"
+          if command -v sha256sum >/dev/null 2>&1; then
+            echo "$checksum  $archive" | sha256sum -c -
+          else
+            echo "$checksum  $archive" | shasum -a 256 -c -
+          fi
+          tar -xzf "$archive"
+          mv antigravity agy
+          chmod +x agy
+          echo "$install_dir" >> "$GITHUB_PATH"
+
+      - name: Install AGY (Windows)
+        if: runner.os == 'Windows'
+        shell: pwsh
+        env:
+          AGY_VERSION: 1.1.13
+          AGY_ARCH: ${{ runner.arch }}
+        run: |
+          $ErrorActionPreference = 'Stop'
+          switch ($env:AGY_ARCH) {
+            'X64' { $archive = 'agy_cli_windows_x64.zip'; $checksum = '9e545832f77adccec08ad48a9db92b31e62c377aab57993946930e0d7d44cb01' }
+            'ARM64' { $archive = 'agy_cli_windows_arm64.zip'; $checksum = 'f9e173c4525178c0d0c4c1381e31debbaa7dc357bd92a97dbced53e74315fbe7' }
+            default { throw "Unsupported AGY architecture: $env:AGY_ARCH" }
+          }
+          $installDir = Join-Path $env:RUNNER_TEMP 'agy'
+          New-Item -ItemType Directory -Force -Path $installDir | Out-Null
+          $archivePath = Join-Path $installDir $archive
+          Invoke-WebRequest "https://github.com/google-antigravity/antigravity-cli/releases/download/$env:AGY_VERSION/$archive" -OutFile $archivePath
+          if ((Get-FileHash $archivePath -Algorithm SHA256).Hash -ne $checksum) {
+            throw 'AGY archive checksum mismatch'
+          }
+          Expand-Archive $archivePath -DestinationPath $installDir -Force
+          Rename-Item (Join-Path $installDir 'antigravity.exe') 'agy.exe'
+          $installDir | Out-File -FilePath $env:GITHUB_PATH -Encoding utf8 -Append
 
       # opencode installs its configured provider's npm package (here
       # @ai-sdk/openai-compatible — see launch::opencode_config) on demand
