+++ docs/providers.md
@@ -179,6 +179,7 @@ installed:
 | `qwen` | Qwen Code | yes |
 | `dsh` | DeepSeek Harness | yes |
 | `hermes` | Hermes Agent | yes, but the daemon holds the key (below) |
+| `agy` | Antigravity CLI (requires `--model`) | yes |
 | `gemini` | Gemini CLI | no: llmman cannot confirm the key would come here rather than go to Google |
 | `cline` | Cline | no: it picks its own model rather than taking llmman's |
 | `kimi` | Kimi Code CLI | no: it picks its own model rather than taking llmman's |
