+++ src/cmd/serve/tests.rs
@@ -9,356 +9,6 @@ use super::sched::{reap_idle_models_once, resolve_keep_alive, DEFAULT_KEEP_ALIVE
 use super::stream::{fold_ollama_lines, stream_ollama};
 use super::*;
 
-#[test]
-fn gemini_terminal_stream_event_preserves_content_and_finish_reason() {
-    let calls = std::cell::RefCell::new(std::collections::BTreeMap::new());
-    let finished = std::cell::Cell::new(false);
-    let event = gemini_sse_line(
-        r#"{"choices":[{"delta":{"content":"done"},"finish_reason":"stop"}],"usage":{"prompt_tokens":3,"completion_tokens":2,"total_tokens":5}}"#,
-        &calls,
-        &finished,
-    );
-
-    let payload = event
-        .strip_prefix("data: ")
-        .and_then(|event| event.strip_suffix("\n\n"))
-        .expect("Gemini SSE event");
-    let response: serde_json::Value = serde_json::from_str(payload).expect("valid JSON event");
-    assert_eq!(
-        response["candidates"][0]["content"]["parts"][0]["text"],
-        "done"
-    );
-    assert_eq!(response["candidates"][0]["finishReason"], "STOP");
-    assert_eq!(
-        response["usageMetadata"],
-        serde_json::json!({
-            "promptTokenCount": 3,
-            "candidatesTokenCount": 2,
-            "totalTokenCount": 5,
-        })
-    );
-}
-
-#[test]
-fn gemini_inline_image_becomes_an_openai_image_part() {
-    let content = gemini_content(&[serde_json::json!({
-        "inlineData": {"mimeType": "image/png", "data": "aGVsbG8="}
-    })])
-    .expect("supported inline image");
-
-    assert_eq!(
-        content,
-        serde_json::json!([{
-            "type": "image_url",
-            "image_url": {"url": "data:image/png;base64,aGVsbG8="}
-        }])
-    );
-}
-
-#[test]
-fn gemini_generation_controls_are_forwarded() {
-    for (seed, presence, frequency, top_k) in [
-        (42, -0.5, 0.75, 40),
-        (0, 0.0, 0.0, 0),
-        (i32::MAX as u64, 1.0, -1.0, u32::MAX),
-    ] {
-        let req: GeminiRequest = serde_json::from_value(serde_json::json!({
-            "contents": [{"role": "user", "parts": [{"text": "hello"}]}],
-            "generationConfig": {
-                "seed": seed,
-                "presencePenalty": presence,
-                "frequencyPenalty": frequency,
-                "topK": top_k
-            }
-        }))
-        .unwrap();
-        let request = gemini_oai_request("selected-model".into(), &req).unwrap();
-        let body = serde_json::to_value(request).unwrap();
-        assert_eq!(body["seed"], seed);
-        assert_eq!(body["presence_penalty"], presence);
-        assert_eq!(body["frequency_penalty"], frequency);
-        assert_eq!(body["top_k"], top_k);
-    }
-}
-
-#[test]
-fn gemini_generation_controls_without_representable_values_remain_unset() {
-    for generation in [
-        serde_json::Value::Null,
-        serde_json::json!({}),
-        serde_json::json!({"presencePenalty": "0.5", "frequencyPenalty": false, "topK": u64::from(u32::MAX) + 1}),
-        serde_json::json!({"topK": 1.5}),
-    ] {
-        let req: GeminiRequest = serde_json::from_value(serde_json::json!({
-            "generationConfig": generation
-        }))
-        .unwrap();
-        let request = gemini_oai_request("selected-model".into(), &req).unwrap();
-        assert_eq!(request.seed, None);
-        assert_eq!(request.presence_penalty, None);
-        assert_eq!(request.frequency_penalty, None);
-        assert_eq!(request.top_k, None);
-    }
-}
-
-#[test]
-fn gemini_tool_config_constrains_backend_tools_to_all_allowed_names() {
-    let config = Some(serde_json::json!({
-        "functionCallingConfig": {"mode": "ANY", "allowedFunctionNames": ["inspect", "search"]}
-    }));
-    let tools = Some(serde_json::json!([{
-        "functionDeclarations": [
-            {"name": "inspect", "parameters": {"type": "object"}},
-            {"name": "search", "parameters": {"type": "object"}},
-            {"name": "delete", "parameters": {"type": "object"}}
-        ]
-    }]));
-
-    assert_eq!(
-        gemini_tool_choice(&config).expect("supported tool config"),
-        Some(serde_json::json!("required"))
-    );
-    assert_eq!(
-        gemini_tools(&tools, &config),
-        Some(serde_json::json!([
-            {"type": "function", "function": {"name": "inspect", "parameters": {"type": "object"}}},
-            {"type": "function", "function": {"name": "search", "parameters": {"type": "object"}}}
-        ]))
-    );
-}
-
-#[test]
-fn gemini_empty_or_omitted_allow_list_keeps_all_tools() {
-    let tools = Some(serde_json::json!([{
-        "functionDeclarations": [{"name": "inspect"}, {"name": "search"}]
-    }]));
-    let expected = Some(serde_json::json!([
-        {"type": "function", "function": {"name": "inspect", "parameters": {"type": "object"}}},
-        {"type": "function", "function": {"name": "search", "parameters": {"type": "object"}}}
-    ]));
-    for config in [
-        serde_json::json!({"functionCallingConfig": {"mode": "ANY"}}),
-        serde_json::json!({"functionCallingConfig": {"mode": "ANY", "allowedFunctionNames": []}}),
-    ] {
-        let config = Some(config);
-        assert_eq!(gemini_tools(&tools, &config), expected);
-        assert_eq!(
-            gemini_tool_choice(&config).unwrap(),
-            Some(serde_json::json!("required"))
-        );
-    }
-}
-
-#[test]
-fn gemini_function_responses_without_ids_match_prior_calls_across_content_blocks() {
-    let req = GeminiRequest {
-        contents: vec![
-            GeminiContent {
-                role: "model".into(),
-                parts: vec![
-                    serde_json::json!({"functionCall": {"name": "lookup", "args": {"id": 1}}}),
-                    serde_json::json!({"functionCall": {"name": "lookup", "args": {"id": 2}}}),
-                ],
-            },
-            GeminiContent {
-                role: "user".into(),
-                parts: vec![
-                    serde_json::json!({"functionResponse": {"name": "lookup", "response": {"id": 1}}}),
-                ],
-            },
-            GeminiContent {
-                role: "user".into(),
-                parts: vec![
-                    serde_json::json!({"functionResponse": {"name": "lookup", "response": {"id": 2}}}),
-                ],
-            },
-        ],
-        system_instruction: None,
-        generation_config: None,
-        tools: None,
-        tool_config: None,
-    };
-
-    let messages = gemini_messages(&req).expect("valid Gemini tool conversation");
-    let calls = messages[0]
-        .tool_calls
-        .as_ref()
-        .expect("assistant tool calls");
-    assert_eq!(
-        messages[1].tool_call_id.as_deref(),
-        Some(calls[0].id.as_str())
-    );
-    assert_eq!(
-        messages[2].tool_call_id.as_deref(),
-        Some(calls[1].id.as_str())
-    );
-}
-
-#[test]
-fn gemini_explicit_function_response_id_is_not_reused_as_a_fallback() {
-    let req = GeminiRequest {
-        contents: vec![
-            GeminiContent {
-                role: "model".into(),
-                parts: vec![
-                    serde_json::json!({"functionCall": {"id": "first", "name": "lookup"}}),
-                    serde_json::json!({"functionCall": {"id": "second", "name": "lookup"}}),
-                ],
-            },
-            GeminiContent {
-                role: "user".into(),
-                parts: vec![
-                    serde_json::json!({"functionResponse": {"id": "first", "name": "lookup"}}),
-                ],
-            },
-            GeminiContent {
-                role: "user".into(),
-                parts: vec![serde_json::json!({"functionResponse": {"name": "lookup"}})],
-            },
-        ],
-        system_instruction: None,
-        generation_config: None,
-        tools: None,
-        tool_config: None,
-    };
-
-    let messages = gemini_messages(&req).expect("valid Gemini tool conversation");
-    assert_eq!(messages[1].tool_call_id.as_deref(), Some("first"));
-    assert_eq!(messages[2].tool_call_id.as_deref(), Some("second"));
-}
-
-#[test]
-fn gemini_generated_ids_do_not_collide_with_supplied_ids() {
-    for explicit_first in [true, false] {
-        let mut parts = vec![
-            serde_json::json!({"functionCall": {"id": "call_lookup_0", "name": "lookup"}}),
-            serde_json::json!({"functionCall": {"name": "lookup"}}),
-        ];
-        if !explicit_first {
-            parts.reverse();
-        }
-        let req: GeminiRequest = serde_json::from_value(serde_json::json!({
-            "contents": [
-                {"role": "model", "parts": parts},
-                {"role": "user", "parts": [
-                    {"functionResponse": {"name": "lookup", "response": {}}},
-                    {"functionResponse": {"name": "lookup", "response": {}}}
-                ]},
-                {"role": "model", "parts": [
-                    {"functionCall": {"id": "call_lookup_1", "name": "lookup"}},
-                    {"functionCall": {"name": "lookup"}}
-                ]},
-                {"role": "user", "parts": [
-                    {"functionResponse": {"name": "lookup", "response": {}}},
-                    {"functionResponse": {"name": "lookup", "response": {}}}
-                ]}
-            ]
-        }))
-        .unwrap();
-        let messages = gemini_messages(&req).unwrap();
-        let ids: Vec<_> = messages
-            .iter()
-            .filter_map(|message| message.tool_calls.as_ref())
-            .flatten()
-            .map(|call| call.id.as_str())
-            .collect();
-        assert_eq!(
-            ids.iter().collect::<std::collections::BTreeSet<_>>().len(),
-            4
-        );
-        assert_eq!(ids[usize::from(!explicit_first)], "call_lookup_0");
-        assert_eq!(ids[2], "call_lookup_1");
-        let response_ids: Vec<_> = messages
-            .iter()
-            .filter_map(|message| message.tool_call_id.as_deref())
-            .collect();
-        assert_eq!(response_ids, ids);
-    }
-}
-
-#[tokio::test]
-async fn gemini_context_overflow_retries_hosted_unless_pinned_local() {
-    let seen = Arc::new(tokio::sync::Mutex::new(Vec::new()));
-    let captured = seen.clone();
-    let app = Router::new().route(
-        "/v1/chat/completions",
-        post(move |Json(body): Json<serde_json::Value>| {
-            let captured = captured.clone();
-            async move {
-                captured.lock().await.push(body.clone());
-                if body["model"] == "gemma4" {
-                    return (
-                        StatusCode::BAD_REQUEST,
-                        r#"{"error":{"message":"request exceeds the available context size","type":"exceed_context_size_error"}}"#,
-                    ).into_response();
-                }
-                (
-                    [("content-type", "text/event-stream")],
-                    "data: {\"choices\":[{\"delta\":{\"content\":\"hosted answer\"},\"finish_reason\":\"stop\"}]}\n\ndata: [DONE]\n\n",
-                ).into_response()
-            }
-        }),
-    );
-    let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-    let port = listener.local_addr().unwrap().port();
-    tokio::spawn(async move { axum::serve(listener, app).await.unwrap() });
-    let state = test_state();
-    let req: GeminiRequest = serde_json::from_value(serde_json::json!({
-        "contents": [{"role": "user", "parts": [{"text": "hello"}]}],
-        "generationConfig": {"temperature": 0.25}
-    }))
-    .unwrap();
-
-    for pinned in [false, true] {
-        seen.lock().await.clear();
-        let headers = if pinned {
-            headers_with(&[("x-llmman-route", "local")])
-        } else {
-            HeaderMap::new()
-        };
-        let resolve = |model: String| {
-            let state = &state;
-            async move {
-                let model = crate::hybrid::local_half(&model).to_string();
-                let target = if model == HOSTED {
-                    remote_target(&format!("http://127.0.0.1:{port}/v1"))
-                } else {
-                    Target::Local(port)
-                };
-                let guard = ActivityGuard::new(state, &model);
-                Ok((model, target, guard))
-            }
-        };
-        let result = with_hybrid_fallback(PAIR, Some(&headers), resolve, |model, target, guard| {
-            gemini_request_to(&state, &req, model, target, guard)
-        })
-        .await;
-        if pinned {
-            assert!(result
-                .unwrap_err()
-                .0
-                .downcast_ref::<ContextOverflow>()
-                .is_some());
-            assert_eq!(seen.lock().await.len(), 1);
-        } else {
-            let response = result.unwrap();
-            assert_eq!(response.status(), StatusCode::OK);
-            let body = axum::body::to_bytes(response.into_body(), usize::MAX)
-                .await
-                .unwrap();
-            assert!(std::str::from_utf8(&body)
-                .unwrap()
-                .contains("hosted answer"));
-            let requests = seen.lock().await;
-            assert_eq!(requests.len(), 2);
-            assert_eq!(requests[0]["model"], "gemma4");
-            assert_eq!(requests[1]["model"], "mock-model");
-            assert_eq!(requests[0]["messages"], requests[1]["messages"]);
-            assert_eq!(requests[1]["temperature"], 0.25);
-        }
-    }
-}
-
 // -- pull/push progress relay -------------------------------------------
 
 /// Regression test for tagless pulls losing their bar: the shim
@@ -421,7 +71,7 @@ fn progress_line_reports_status_only_snapshots_and_clamps_completed() {
 
 // -- request targets (local backend vs remote provider) -----------------
 
-fn remote_target(base_url: &str) -> Target {
+pub(super) fn remote_target(base_url: &str) -> Target {
     remote_target_on(base_url, Wire::OpenAi)
 }
 
@@ -1562,7 +1212,7 @@ async fn an_unload_is_forwarded_to_every_peer() {
 
 // -- hybrid pairs (one reference, a local and a hosted half) -------------
 
-fn headers_with(pairs: &[(&str, &str)]) -> HeaderMap {
+pub(super) fn headers_with(pairs: &[(&str, &str)]) -> HeaderMap {
     let mut headers = HeaderMap::new();
     for (name, value) in pairs {
         headers.insert(
@@ -1577,8 +1227,8 @@ fn pair(reference: &'static str) -> crate::hybrid::Pair<'static> {
     crate::hybrid::split_ref(reference).expect("test reference must be a pair")
 }
 
-const PAIR: &str = "llmman.hybrid/gemma4,anthropic/claude-sonnet-4-5";
-const HOSTED: &str = "llmman.provider/anthropic/claude-sonnet-4-5";
+pub(super) const PAIR: &str = "llmman.hybrid/gemma4,anthropic/claude-sonnet-4-5";
+pub(super) const HOSTED: &str = "llmman.provider/anthropic/claude-sonnet-4-5";
 
 /// What comes back is one half's own ordinary reference, with
 /// nothing left for anything downstream to special-case.
@@ -2330,7 +1980,7 @@ async fn a_configured_providers_models_are_asked_of_its_endpoint() {
 
 // -- Idle-timeout auto-unload reaper --------------------------------------
 
-fn test_state() -> AppState {
+pub(super) fn test_state() -> AppState {
     test_state_at(std::env::temp_dir())
 }
 
@@ -5989,89 +5639,4 @@ async fn an_unenforced_policy_strips_its_keys_without_refusing_anyone() {
         client_api_key(Some(&seen[1])).as_deref(),
         Some("sk-provider")
     );
-}
-
-#[test]
-fn gemini_invalid_or_unsupported_seeds_return_bad_request() {
-    for seed in [
-        serde_json::json!(-1),
-        serde_json::json!(i32::MIN),
-        serde_json::json!(i64::from(i32::MIN) - 1),
-        serde_json::json!(i64::from(i32::MAX) + 1),
-        serde_json::json!(u64::MAX),
-        serde_json::json!(1.5),
-        serde_json::json!("42"),
-        serde_json::json!(true),
-        serde_json::Value::Null,
-    ] {
-        let req: GeminiRequest = serde_json::from_value(serde_json::json!({
-            "generationConfig": {"seed": seed}
-        }))
-        .unwrap();
-        let error = gemini_oai_request("model".into(), &req).unwrap_err();
-        assert_eq!(error.1, StatusCode::BAD_REQUEST);
-        assert!(error.0.to_string().contains("seed"));
-    }
-}
-
-#[test]
-fn gemini_stream_relays_provider_errors_after_partial_content() {
-    let calls = std::cell::RefCell::new(std::collections::BTreeMap::new());
-    let finished = std::cell::Cell::new(false);
-    assert!(!gemini_sse_line(
-        r#"{"choices":[{"delta":{"content":"partial"}}]}"#,
-        &calls,
-        &finished
-    )
-    .is_empty());
-    for event in [
-        serde_json::json!({"error": {"message": "upstream failed", "type": "api_error"}}),
-        serde_json::json!({"error": "upstream failed"}),
-    ] {
-        let output = gemini_sse_line(&event.to_string(), &calls, &finished);
-        let actual: serde_json::Value =
-            serde_json::from_str(output.strip_prefix("data: ").unwrap().trim()).unwrap();
-        assert_eq!(actual, event);
-    }
-}
-
-#[test]
-fn gemini_tool_schema_normalizes_nested_types_without_changing_data() {
-    let schema = serde_json::json!({
-        "type": "OBJECT", "required": ["type"],
-        "properties": {
-            "type": {"type": "STRING", "enum": ["OBJECT", "STRING"]},
-            "values": {"type": "ARRAY", "items": {"type": "OBJECT", "properties": {
-                "count": {"type": "INTEGER"}, "score": {"type": "NUMBER"},
-                "enabled": {"type": "BOOLEAN"}}}},
-            "optional": {"anyOf": [{"type": "STRING"}, {"type": "NULL"}]}
-        },
-        "default": {"type": "OBJECT"}, "example": {"type": "STRING"}
-    });
-    let tools = Some(serde_json::json!([{"functionDeclarations": [
-        {"name": "proto", "parameters": schema},
-        {"name": "json", "parametersJsonSchema": schema}
-    ]}]));
-    let result = gemini_tools(&tools, &None).unwrap();
-    let normalized = &result[0]["function"]["parameters"];
-    assert_eq!(normalized["type"], "object");
-    assert_eq!(normalized["properties"]["type"]["type"], "string");
-    assert_eq!(
-        normalized["properties"]["type"]["enum"],
-        schema["properties"]["type"]["enum"]
-    );
-    assert_eq!(normalized["properties"]["values"]["type"], "array");
-    let items = &normalized["properties"]["values"]["items"];
-    assert_eq!(items["type"], "object");
-    assert_eq!(items["properties"]["count"]["type"], "integer");
-    assert_eq!(items["properties"]["score"]["type"], "number");
-    assert_eq!(items["properties"]["enabled"]["type"], "boolean");
-    assert_eq!(
-        normalized["properties"]["optional"]["anyOf"][1]["type"],
-        "null"
-    );
-    assert_eq!(normalized["default"], schema["default"]);
-    assert_eq!(normalized["example"], schema["example"]);
-    assert_eq!(normalized["required"], schema["required"]);
-    assert_eq!(result[1]["function"]["parameters"], schema);
 }
